// tiffin/src/models/api/photo/choiceSearch/request.ts
export type PhotoChoiceSearchRequest = {
  master_prefecture_id?: number;
  master_area_id?: number;
  master_detailed_area_id?: number;
  master_railroad_station_id?: number;
  master_genre_id?: number;
  master_detailed_genre_id?: number;
  master_cooking_id?: number;
  price_min?: number;
  price_max?: number;
};